Serper wrapper class in a separate module.
